--- godeltaprof/internal/pprof/mapping.go
@@ -1,176 +1,0 @@
-package pprof
-
-import (
-	"bytes"
-	"os"
-	"strconv"
-	"strings"
-)
-
-type memMap struct {
-	// initialized as reading mapping
-	start   uintptr // Address at which the binary (or DLL) is loaded into memory.
-	end     uintptr // The limit of the address range occupied by this mapping.
-	offset  uint64  // offset in the binary that corresponds to the first mapped address.
-	file    string  // The object this entry is loaded from.
-	buildID string  // A string that uniquely identifies a particular program version with high probability.
-
-	funcs symbolizeFlag
-	fake  bool // map entry was faked; /proc/self/maps wasn't available
-}
-
-// symbolizeFlag keeps track of symbolization result.
-//
-//	0                  : no symbol lookup was performed
-//	1<<0 (lookupTried) : symbol lookup was performed
-//	1<<1 (lookupFailed): symbol lookup was performed but failed
-type symbolizeFlag uint8
-
-const (
-	lookupTried  symbolizeFlag = 1 << iota
-	lookupFailed symbolizeFlag = 1 << iota
-)
-
-func ReadMapping() []memMap {
-	data, _ := os.ReadFile("/proc/self/maps")
-	var mem []memMap
-	parseProcSelfMaps(data, func(lo, hi, offset uint64, file, buildID string) {
-		mem = append(mem, memMap{
-			start:   uintptr(lo),
-			end:     uintptr(hi),
-			offset:  offset,
-			file:    file,
-			buildID: buildID,
-			fake:    false,
-		})
-	})
-	if len(mem) == 0 { // pprof expects a map entry, so fake one.
-		mem = []memMap{{
-			start:   uintptr(0),
-			end:     uintptr(0),
-			offset:  0,
-			file:    "",
-			buildID: "",
-			fake:    true,
-		}}
-	}
-	return mem
-}
-
-var space = []byte(" ")
-var newline = []byte("\n")
-
-func parseProcSelfMaps(data []byte, addMapping func(lo, hi, offset uint64, file, buildID string)) {
-	// $ cat /proc/self/maps
-	// 00400000-0040b000 r-xp 00000000 fc:01 787766                             /bin/cat
-	// 0060a000-0060b000 r--p 0000a000 fc:01 787766                             /bin/cat
-	// 0060b000-0060c000 rw-p 0000b000 fc:01 787766                             /bin/cat
-	// 014ab000-014cc000 rw-p 00000000 00:00 0                                  [heap]
-	// 7f7d76af8000-7f7d7797c000 r--p 00000000 fc:01 1318064                    /usr/lib/locale/locale-archive
-	// 7f7d7797c000-7f7d77b36000 r-xp 00000000 fc:01 1180226                    /lib/x86_64-linux-gnu/libc-2.19.so
-	// 7f7d77b36000-7f7d77d36000 ---p 001ba000 fc:01 1180226                    /lib/x86_64-linux-gnu/libc-2.19.so
-	// 7f7d77d36000-7f7d77d3a000 r--p 001ba000 fc:01 1180226                    /lib/x86_64-linux-gnu/libc-2.19.so
-	// 7f7d77d3a000-7f7d77d3c000 rw-p 001be000 fc:01 1180226                    /lib/x86_64-linux-gnu/libc-2.19.so
-	// 7f7d77d3c000-7f7d77d41000 rw-p 00000000 00:00 0
-	// 7f7d77d41000-7f7d77d64000 r-xp 00000000 fc:01 1180217                    /lib/x86_64-linux-gnu/ld-2.19.so
-	// 7f7d77f3f000-7f7d77f42000 rw-p 00000000 00:00 0
-	// 7f7d77f61000-7f7d77f63000 rw-p 00000000 00:00 0
-	// 7f7d77f63000-7f7d77f64000 r--p 00022000 fc:01 1180217                    /lib/x86_64-linux-gnu/ld-2.19.so
-	// 7f7d77f64000-7f7d77f65000 rw-p 00023000 fc:01 1180217                    /lib/x86_64-linux-gnu/ld-2.19.so
-	// 7f7d77f65000-7f7d77f66000 rw-p 00000000 00:00 0
-	// 7ffc342a2000-7ffc342c3000 rw-p 00000000 00:00 0                          [stack]
-	// 7ffc34343000-7ffc34345000 r-xp 00000000 00:00 0                          [vdso]
-	// ffffffffff600000-ffffffffff601000 r-xp 00000000 00:00 0                  [vsyscall]
-
-	var line []byte
-	// next removes and returns the next field in the line.
-	// It also removes from line any spaces following the field.
-	next := func() []byte {
-		var f []byte
-		f, line, _ = bytesCut(line, space)
-		line = bytes.TrimLeft(line, " ")
-		return f
-	}
-
-	for len(data) > 0 {
-		line, data, _ = bytesCut(data, newline)
-		addr := next()
-		loStr, hiStr, ok := stringsCut(string(addr), "-")
-		if !ok {
-			continue
-		}
-		lo, err := strconv.ParseUint(loStr, 16, 64)
-		if err != nil {
-			continue
-		}
-		hi, err := strconv.ParseUint(hiStr, 16, 64)
-		if err != nil {
-			continue
-		}
-		perm := next()
-		if len(perm) < 4 || perm[2] != 'x' {
-			// Only interested in executable mappings.
-			continue
-		}
-		offset, err := strconv.ParseUint(string(next()), 16, 64)
-		if err != nil {
-			continue
-		}
-		next()          // dev
-		inode := next() // inode
-		if line == nil {
-			continue
-		}
-		file := string(line)
-
-		// Trim deleted file marker.
-		deletedStr := " (deleted)"
-		deletedLen := len(deletedStr)
-		if len(file) >= deletedLen && file[len(file)-deletedLen:] == deletedStr {
-			file = file[:len(file)-deletedLen]
-		}
-
-		if len(inode) == 1 && inode[0] == '0' && file == "" {
-			// Huge-page text mappings list the initial fragment of
-			// mapped but unpopulated memory as being inode 0.
-			// Don't report that part.
-			// But [vdso] and [vsyscall] are inode 0, so let non-empty file names through.
-			continue
-		}
-
-		// TODO: pprof's remapMappingIDs makes one adjustment:
-		// 1. If there is an /anon_hugepage mapping first and it is
-		// consecutive to a next mapping, drop the /anon_hugepage.
-		// There's no indication why this is needed.
-		// Let's try not doing this and see what breaks.
-		// If we do need it, it would go here, before we
-		// enter the mappings into b.mem in the first place.
-
-		buildID, _ := elfBuildID(file)
-		addMapping(lo, hi, offset, file, buildID)
-	}
-}
-
-// Cut slices s around the first instance of sep,
-// returning the text before and after sep.
-// The found result reports whether sep appears in s.
-// If sep does not appear in s, cut returns s, nil, false.
-//
-// Cut returns slices of the original slice s, not copies.
-func bytesCut(s, sep []byte) (before, after []byte, found bool) {
-	if i := bytes.Index(s, sep); i >= 0 {
-		return s[:i], s[i+len(sep):], true
-	}
-	return s, nil, false
-}
-
-// Cut slices s around the first instance of sep,
-// returning the text before and after sep.
-// The found result reports whether sep appears in s.
-// If sep does not appear in s, cut returns s, "", false.
-func stringsCut(s, sep string) (before, after string, found bool) {
-	if i := strings.Index(s, sep); i >= 0 {
-		return s[:i], s[i+len(sep):], true
-	}
-	return s, "", false
-}
